strict type fix

+++ src/Bridges/SecurityHttp/SessionStorage.php
@@ -84,7 +84,7 @@ final class SessionStorage implements Nette\Security\UserStorage
 
 	public function setExpiration(?string $time, bool $clearIdentity = false): void
 	{
-		$this->expireTime = $time ? Nette\Utils\DateTime::from($time)->format('U') : null;
+		$this->expireTime = $time ? (int) Nette\Utils\DateTime::from($time)->format('U') : null;
 		$this->expireIdentity = $clearIdentity;
 
 		if ($this->sessionSection && $this->sessionSection->get('authenticated')) {
